move platforms customizations to traits
