+++ src/Persistence/Sql/Connection.php
@@ -15,7 +15,6 @@ use Doctrine\DBAL\Platforms\OraclePlatform;
 use Doctrine\DBAL\Platforms\PostgreSQL94Platform;
 use Doctrine\DBAL\Platforms\SQLServer2012Platform;
 use Doctrine\DBAL\Result as DbalResult;
-use Doctrine\DBAL\Schema\Sequence;
 
 /**
  * Class for establishing and maintaining connection with your database.
@@ -274,312 +273,18 @@ abstract class Connection
             }, null, DbalConnection::class)();
         }
 
-
-
-
-
-
         if ($dbalConnection->getDatabasePlatform() instanceof PostgreSQL94Platform) {
             \Closure::bind(function () use ($dbalConnection) {
                 $dbalConnection->platform = new class() extends PostgreSQL94Platform {
-                    // PostgreSQL DBAL platform uses SERIAL column type for autoincrement which does not increment
-                    // when a row with a not-null PK is inserted like Sqlite or MySQL does, unify the behaviour
-
-
-
-                    // TODO probably there is a better place to fix
-                    protected function _getCreateTableSQL($name, array $columns, array $options = [])
-                    {
-                        $sqls = parent::_getCreateTableSQL($name, $columns, $options);
-
-                        $conn = new Postgresql\Connection();
-
-
-
-                        $pkColName = null;
-                        foreach ($columns as $c) {
-                            if ($c['autoincrement']) {
-                                $pkColName = trim($c['name'], '"');
-                            }
-                        }
-                        if ($pkColName === null) {
-                            debug_print_backtrace(\DEBUG_BACKTRACE_IGNORE_ARGS);
-                            return $sqls;
-                        }
-
-
-                        $t = trim($name, '"');
-                        $pkseq = $t . '_' . $pkColName . '_seq';
-
-                        $sqls[] = $conn->expr(
-                            // else branch should be maybe (because of concurrency) put into after update trigger
-                            // with pure nextval instead of setval with a loop like in Oracle trigger
-                            str_replace('[pk_seq]', '\'' . $pkseq . '\'', <<<'EOF'
-                                CREATE OR REPLACE FUNCTION {trigger_func}()
-                                RETURNS trigger AS $$
-                                DECLARE
-                                    atk4__pk_seq_last__ {table}.{pk}%TYPE;
-                                BEGIN
-                                    IF (NEW.{pk} IS NULL) THEN
-                                        NEW.{pk} := nextval([pk_seq]);
-                                    ELSE
-                                        SELECT COALESCE(last_value, 0) INTO atk4__pk_seq_last__ FROM {pk_seq};
-                                        IF (atk4__pk_seq_last__ <= NEW.{pk}) THEN
-                                            atk4__pk_seq_last__  := setval([pk_seq], NEW.{pk}, true);
-                                        END IF;
-                                    END IF;
-                                    RETURN NEW;
-                                END;
-                                $$ LANGUAGE plpgsql;
-                                EOF),
-                            [
-                                'table' => $t,
-                                'pk' => $pkColName,
-                                'pk_seq' => $pkseq,
-                                'trigger_func' => $t . '_func',
-                            ]
-                            // TODO should not exist... (no OR REPLACE) also for Oracle
-                        )->render();
-                        // function is not dropped when table is, we should use only one
-
-
-                        $sqls[] = $conn->expr(
-                            str_replace('[pk_seq]', '\'' . $pkseq . '\'', <<<'EOF'
-                                CREATE TRIGGER {trigger}
-                                BEFORE INSERT OR UPDATE
-                                ON {table}
-                                FOR EACH ROW
-                                EXECUTE PROCEDURE {trigger_func}();
-                                EOF),
-                            [
-                                'table' => $t,
-                                'trigger' => $t . '_tri',
-                                'trigger_func' => $t . '_func',
-                            ]
-                        )->render();
-                        // TODO procedure name/fucntion
-                        // TODO test if it is deleteted when the table it
-
-
-                        return $sqls;
-                    }
-
-
-
-
-                    /*
-                    public function getCreateAutoincrementSql($name, $table, $start = 1)
-                    {
-                        $sqls = parent::getCreateAutoincrementSql($name, $table, $start);
-
-                        // replace https://github.com/doctrine/dbal/blob/3.1.3/src/Platforms/OraclePlatform.php#L526-L546
-                        $tableIdentifier = \Closure::bind(fn () => $this->normalizeIdentifier($table), $this, OraclePlatform::class)();
-                        $nameIdentifier = \Closure::bind(fn () => $this->normalizeIdentifier($name), $this, OraclePlatform::class)();
-                        $aiTriggerName = \Closure::bind(fn () => $this->getAutoincrementIdentifierName($tableIdentifier), $this, OraclePlatform::class)();
-                        $aiSequenceName = $this->getIdentitySequenceName($tableIdentifier->getQuotedName($this), $nameIdentifier->getQuotedName($this));
-                        assert(str_starts_with($sqls[count($sqls) - 1], 'CREATE TRIGGER ' . $aiTriggerName . "\n"));
-
-                        $conn = new Oracle\Connection();
-                        $pkSeq = \Closure::bind(fn () => $this->normalizeIdentifier($aiSequenceName), $this, OraclePlatform::class)()->getName();
-                        $sqls[count($sqls) - 1] = $conn->expr(
-                            str_replace('[pk_seq]', '\'' . $pkSeq . '\'', <<<'EOT'
-                                CREATE OR REPLACE TRIGGER {trigger}
-                                    BEFORE INSERT OR UPDATE
-                                    ON {table}
-                                    FOR EACH ROW
-                                DECLARE
-                                    pk_seq_last {table}.{pk}%TYPE;
-                                BEGIN
-                                    IF (NVL(:NEW.{pk}, 0) = 0) THEN
-                                        SELECT {pk_seq}.NEXTVAL INTO :NEW.{pk} FROM DUAL;
-                                    ELSE
-                                        SELECT NVL(LAST_NUMBER, 0) INTO pk_seq_last FROM USER_SEQUENCES WHERE SEQUENCE_NAME = [pk_seq];
-                                        WHILE pk_seq_last <= :NEW.{pk}
-                                        LOOP
-                                            SELECT {pk_seq}.NEXTVAL + 1 INTO pk_seq_last FROM DUAL;
-                                        END LOOP;
-                                    END IF;
-                                END;
-                                EOT),
-                            [
-                                'trigger' => \Closure::bind(fn () => $this->normalizeIdentifier($aiTriggerName), $this, OraclePlatform::class)()->getName(),
-                                'table' => $tableIdentifier->getName(),
-                                'pk' => $nameIdentifier->getName(),
-                                'pk_seq' => $pkSeq,
-                            ]
-                        )->render();
-
-                        return $sqls;
-                    }*/
+                    use Postgresql\PlatformTrait;
                 };
             }, null, DbalConnection::class)();
         }
 
-
-
-
-
-
-
-
         if ($dbalConnection->getDatabasePlatform() instanceof SQLServer2012Platform) {
             \Closure::bind(function () use ($dbalConnection) {
                 $dbalConnection->platform = new class() extends SQLServer2012Platform {
-                    // SQL Server DBAL platform has buggy identifier escaping, fix until fixed officially, see:
-                    // https://github.com/doctrine/dbal/pull/4360
-
-                    protected function getCreateColumnCommentSQL($tableName, $columnName, $comment)
-                    {
-                        if (strpos($tableName, '.') !== false) {
-                            [$schemaName, $tableName] = explode('.', $tableName, 2);
-                        } else {
-                            $schemaName = $this->getDefaultSchemaName();
-                        }
-
-                        return $this->getAddExtendedPropertySQL(
-                            'MS_Description',
-                            (string) $comment,
-                            'SCHEMA',
-                            $schemaName,
-                            'TABLE',
-                            $tableName,
-                            'COLUMN',
-                            $columnName
-                        );
-                    }
-
-                    protected function getAlterColumnCommentSQL($tableName, $columnName, $comment)
-                    {
-                        if (strpos($tableName, '.') !== false) {
-                            [$schemaName, $tableName] = explode('.', $tableName, 2);
-                        } else {
-                            $schemaName = $this->getDefaultSchemaName();
-                        }
-
-                        return $this->getUpdateExtendedPropertySQL(
-                            'MS_Description',
-                            (string) $comment,
-                            'SCHEMA',
-                            $schemaName,
-                            'TABLE',
-                            $tableName,
-                            'COLUMN',
-                            $columnName
-                        );
-                    }
-
-                    protected function getDropColumnCommentSQL($tableName, $columnName)
-                    {
-                        if (strpos($tableName, '.') !== false) {
-                            [$schemaName, $tableName] = explode('.', $tableName, 2);
-                        } else {
-                            $schemaName = $this->getDefaultSchemaName();
-                        }
-
-                        return $this->getDropExtendedPropertySQL(
-                            'MS_Description',
-                            'SCHEMA',
-                            $schemaName,
-                            'TABLE',
-                            $tableName,
-                            'COLUMN',
-                            $columnName
-                        );
-                    }
-
-                    private function quoteSingleIdentifierAsStringLiteral(string $levelName): string
-                    {
-                        return $this->quoteStringLiteral(preg_replace('~^\[|\]$~s', '', $levelName));
-                    }
-
-                    public function getAddExtendedPropertySQL(
-                        $name,
-                        $value = null,
-                        $level0Type = null,
-                        $level0Name = null,
-                        $level1Type = null,
-                        $level1Name = null,
-                        $level2Type = null,
-                        $level2Name = null
-                    ) {
-                        return 'EXEC sp_addextendedproperty'
-                            . ' N' . $this->quoteStringLiteral($name) . ', N' . $this->quoteStringLiteral((string) $value)
-                            . ', N' . $this->quoteStringLiteral((string) $level0Type)
-                            . ', ' . $this->quoteSingleIdentifierAsStringLiteral((string) $level0Name)
-                            . ', N' . $this->quoteStringLiteral((string) $level1Type)
-                            . ', ' . $this->quoteSingleIdentifierAsStringLiteral((string) $level1Name)
-                            . (
-                                $level2Type !== null || $level2Name !== null
-                                ? ', N' . $this->quoteStringLiteral((string) $level2Type)
-                                  . ', ' . $this->quoteSingleIdentifierAsStringLiteral((string) $level2Name)
-                                : ''
-                            );
-                    }
-
-                    public function getDropExtendedPropertySQL(
-                        $name,
-                        $level0Type = null,
-                        $level0Name = null,
-                        $level1Type = null,
-                        $level1Name = null,
-                        $level2Type = null,
-                        $level2Name = null
-                    ) {
-                        return 'EXEC sp_dropextendedproperty'
-                            . ' N' . $this->quoteStringLiteral($name)
-                            . ', N' . $this->quoteStringLiteral((string) $level0Type)
-                            . ', ' . $this->quoteSingleIdentifierAsStringLiteral((string) $level0Name)
-                            . ', N' . $this->quoteStringLiteral((string) $level1Type)
-                            . ', ' . $this->quoteSingleIdentifierAsStringLiteral((string) $level1Name)
-                            . (
-                                $level2Type !== null || $level2Name !== null
-                                ? ', N' . $this->quoteStringLiteral((string) $level2Type)
-                                  . ', ' . $this->quoteSingleIdentifierAsStringLiteral((string) $level2Name)
-                                : ''
-                            );
-                    }
-
-                    public function getUpdateExtendedPropertySQL(
-                        $name,
-                        $value = null,
-                        $level0Type = null,
-                        $level0Name = null,
-                        $level1Type = null,
-                        $level1Name = null,
-                        $level2Type = null,
-                        $level2Name = null
-                    ) {
-                        return 'EXEC sp_updateextendedproperty'
-                            . ' N' . $this->quoteStringLiteral($name) . ', N' . $this->quoteStringLiteral((string) $value)
-                            . ', N' . $this->quoteStringLiteral((string) $level0Type)
-                            . ', ' . $this->quoteSingleIdentifierAsStringLiteral((string) $level0Name)
-                            . ', N' . $this->quoteStringLiteral((string) $level1Type)
-                            . ', ' . $this->quoteSingleIdentifierAsStringLiteral((string) $level1Name)
-                            . (
-                                $level2Type !== null || $level2Name !== null
-                                ? ', N' . $this->quoteStringLiteral((string) $level2Type)
-                                  . ', ' . $this->quoteSingleIdentifierAsStringLiteral((string) $level2Name)
-                                : ''
-                            );
-                    }
-
-                    protected function getCommentOnTableSQL(string $tableName, ?string $comment): string
-                    {
-                        if (strpos($tableName, '.') !== false) {
-                            [$schemaName, $tableName] = explode('.', $tableName, 2);
-                        } else {
-                            $schemaName = $this->getDefaultSchemaName();
-                        }
-
-                        return $this->getAddExtendedPropertySQL(
-                            'MS_Description',
-                            (string) $comment,
-                            'SCHEMA',
-                            $schemaName,
-                            'TABLE',
-                            $tableName
-                        );
-                    }
+                    use Mssql\PlatformTrait;
                 };
             }, null, DbalConnection::class)();
         }
@@ -587,91 +292,7 @@ abstract class Connection
         if ($dbalConnection->getDatabasePlatform() instanceof OraclePlatform) {
             \Closure::bind(function () use ($dbalConnection) {
                 $dbalConnection->platform = new class() extends OraclePlatform {
-                    // Oracle CLOB/BLOB has limited SQL support, see:
-                    // https://stackoverflow.com/questions/12980038/ora-00932-inconsistent-datatypes-expected-got-clob#12980560
-                    // fix this Oracle inconsistency by using VARCHAR/VARBINARY instead (but limited to 4000 bytes)
-
-                    private function forwardTypeDeclarationSQL(string $targetMethodName, array $column): string
-                    {
-                        $backtrace = debug_backtrace(\DEBUG_BACKTRACE_PROVIDE_OBJECT | \DEBUG_BACKTRACE_IGNORE_ARGS);
-                        foreach ($backtrace as $frame) {
-                            if ($this === ($frame['object'] ?? null)
-                                && $targetMethodName === ($frame['function'] ?? null)) {
-                                throw new Exception('Long CLOB/TEXT (4000+ bytes) is not supported for Oracle');
-                            }
-                        }
-
-                        return $this->{$targetMethodName}($column);
-                    }
-
-                    public function getClobTypeDeclarationSQL(array $column)
-                    {
-                        $column['length'] = $this->getVarcharMaxLength();
-
-                        return $this->forwardTypeDeclarationSQL('getVarcharTypeDeclarationSQL', $column);
-                    }
-
-                    public function getBlobTypeDeclarationSQL(array $column)
-                    {
-                        $column['length'] = $this->getBinaryMaxLength();
-
-                        return $this->forwardTypeDeclarationSQL('getBinaryTypeDeclarationSQL', $column);
-                    }
-
-                    // Oracle DBAL platform autoincrement implementation does not increment like
-                    // Sqlite or MySQL does, unify the behaviour
-
-                    public function getCreateSequenceSQL(Sequence $sequence)
-                    {
-                        $sequence->setCache(1);
-
-                        return parent::getCreateSequenceSQL($sequence);
-                    }
-
-                    public function getCreateAutoincrementSql($name, $table, $start = 1)
-                    {
-                        $sqls = parent::getCreateAutoincrementSql($name, $table, $start);
-
-                        // replace https://github.com/doctrine/dbal/blob/3.1.3/src/Platforms/OraclePlatform.php#L526-L546
-                        $tableIdentifier = \Closure::bind(fn () => $this->normalizeIdentifier($table), $this, OraclePlatform::class)();
-                        $nameIdentifier = \Closure::bind(fn () => $this->normalizeIdentifier($name), $this, OraclePlatform::class)();
-                        $aiTriggerName = \Closure::bind(fn () => $this->getAutoincrementIdentifierName($tableIdentifier), $this, OraclePlatform::class)();
-                        $aiSequenceName = $this->getIdentitySequenceName($tableIdentifier->getQuotedName($this), $nameIdentifier->getQuotedName($this));
-                        assert(str_starts_with($sqls[count($sqls) - 1], 'CREATE TRIGGER ' . $aiTriggerName . "\n"));
-
-                        $conn = new Oracle\Connection();
-                        $pkSeq = \Closure::bind(fn () => $this->normalizeIdentifier($aiSequenceName), $this, OraclePlatform::class)()->getName();
-                        $sqls[count($sqls) - 1] = $conn->expr(
-                            // else branch should be maybe (because of concurrency) put into after update trigger
-                            str_replace('[pk_seq]', '\'' . $pkSeq . '\'', <<<'EOT'
-                                CREATE OR REPLACE TRIGGER {trigger}
-                                    BEFORE INSERT OR UPDATE
-                                    ON {table}
-                                    FOR EACH ROW
-                                DECLARE
-                                    atk4__pk_seq_last__ {table}.{pk}%TYPE;
-                                BEGIN
-                                    IF (:NEW.{pk} IS NULL) THEN
-                                        SELECT {pk_seq}.NEXTVAL INTO :NEW.{pk} FROM DUAL;
-                                    ELSE
-                                        SELECT LAST_NUMBER INTO atk4__pk_seq_last__ FROM USER_SEQUENCES WHERE SEQUENCE_NAME = [pk_seq];
-                                        WHILE atk4__pk_seq_last__ <= :NEW.{pk}
-                                        LOOP
-                                            SELECT {pk_seq}.NEXTVAL + 1 INTO atk4__pk_seq_last__ FROM DUAL;
-                                        END LOOP;
-                                    END IF;
-                                END;
-                                EOT),
-                            [
-                                'trigger' => \Closure::bind(fn () => $this->normalizeIdentifier($aiTriggerName), $this, OraclePlatform::class)()->getName(),
-                                'table' => $tableIdentifier->getName(),
-                                'pk' => $nameIdentifier->getName(),
-                                'pk_seq' => $pkSeq,
-                            ]
-                        )->render();
-
-                        return $sqls;
-                    }
+                    use Oracle\PlatformTrait;
                 };
             }, null, DbalConnection::class)();
         }
